update button logic
